Refactor to remove circular dependency
