+++ app/webpack/users/edit/ducks/relationships.js
@@ -1,8 +1,6 @@
 import _ from "lodash";
 import inatjs from "inaturalistjs";
 
-import { fetchUserSettings } from "./user_settings";
-
 const SET_RELATIONSHIPS = "user/edit/SET_RELATIONSHIPS";
 const SET_RELATIONSHIP_TO_DELETE = "user/edit/SET_RELATIONSHIP_TO_DELETE";
 const SET_USER_AUTOCOMPLETE = "user/edit/SET_USER_AUTOCOMPLETE";
@@ -261,40 +259,4 @@ export function deleteRelationship( ) {
       dispatch( fetchRelationships( ) );
     } ).catch( e => console.log( `Failed to delete relationships: ${e}` ) );
   };
-}
-
-export function muteUser( id ) {
-  const params = { id };
-  return dispatch => inatjs.users.mute( params ).then( ( ) => {
-    dispatch( fetchUserSettings( false, true ) );
-  } ).catch( e => console.log( `Failed to mute user: ${e}` ) );
-}
-
-export function unmuteUser( id ) {
-  const params = { id };
-  return dispatch => inatjs.users.unmute( params ).then( ( ) => {
-    dispatch( fetchUserSettings( false, true ) );
-  } ).catch( e => console.log( `Failed to unmute user: ${e}` ) );
-}
-
-export function blockUser( id ) {
-  const params = { id };
-  return dispatch => inatjs.users.block( params ).then( ( ) => {
-    dispatch( fetchUserSettings( false, true ) );
-  } ).catch( e => {
-    e.response.json( ).then( json => {
-      const baseErrors = _.get( json, "error.original.errors.base", [] );
-      if ( baseErrors.length > 0 ) {
-        alert( baseErrors.join( "; " ) );
-      }
-    } );
-    console.log( `Failed to block user: ${e}` );
-  } );
-}
-
-export function unblockUser( id ) {
-  const params = { id };
-  return dispatch => inatjs.users.unblock( params ).then( ( ) => {
-    dispatch( fetchUserSettings( false, true ) );
-  } ).catch( e => console.log( `Failed to unblock user: ${e}` ) );
 }
